task_2.3.1_v2

+++ src/main/java/org/example/crud/dao/UserDao.java
@@ -12,7 +12,7 @@ public interface UserDao {
 
     void save(User user);
 
-    void update(int id, User user);
+    void update(long id, User user);
 
-    void delete(int id);
+    void delete(long id);
 }
